Setup controller interface for docs
