+++ ldes-server-domain/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/domain/view/valueobject/DcatView.java
@@ -7,18 +7,18 @@ import static org.apache.commons.lang3.Validate.notNull;
 
 public class DcatView {
 
-    private final ViewName viewName;
-    private final Model dcat;
+	private final ViewName viewName;
+	private final Model dcat;
 
-    private DcatView(ViewName viewName, Model dcat) {
-        this.viewName = viewName;
-        this.dcat = dcat;
-    }
+	private DcatView(ViewName viewName, Model dcat) {
+		this.viewName = viewName;
+		this.dcat = dcat;
+	}
 
-    public static DcatView from(ViewName viewName, Model dcat) {
-        return new DcatView(notNull(viewName), dcat);
-    }
+	public static DcatView from(ViewName viewName, Model dcat) {
+		return new DcatView(notNull(viewName), dcat);
+	}
 
-    // TODO TVB: 24/05/2023 eq and hash
+	// TODO TVB: 24/05/2023 eq and hash
 
 }
