+++ ldes-server-infra-mongo/mongo-repository/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/infra/mongo/view/entity/DataServiceEntity.java
@@ -6,22 +6,22 @@ import org.springframework.data.mongodb.core.mapping.Document;
 @Document(collection = "dcat_dataservice")
 public class DataServiceEntity {
 
-    @Id
-    private final String viewName;
+	@Id
+	private final String viewName;
 
-    private final String model;
+	private final String model;
 
-    public DataServiceEntity(String viewName, String model) {
-        this.viewName = viewName;
-        this.model = model;
-    }
+	public DataServiceEntity(String viewName, String model) {
+		this.viewName = viewName;
+		this.model = model;
+	}
 
-    public String getViewName() {
-        return viewName;
-    }
+	public String getViewName() {
+		return viewName;
+	}
 
-    public String getModel() {
-        return model;
-    }
+	public String getModel() {
+		return model;
+	}
 
 }
